requirements.txt
haystack-ai
qdrant-haystack
docling
docling-haystack
google-genai-haystack
